requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
# audioop-lts==0.2.2
beartype==0.21.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
psycopg2-binary
bs4==0.0.2
cachelib==0.13.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
coloredlogs==15.0.1
confection==0.1.5
cramjam==2.11.0
cryptography==3.4.7
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
ddgs==9.5.2
decorator==5.2.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docx==0.2.4
dulwich==0.24.1
easygui==0.98.3
et_xmlfile==2.0.0
executing==2.2.1
ezdxf==0.16.3
fake-useragent==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastcore==1.8.7
fastdownload==0.0.7
fastprogress==1.0.3
fasttransform==0.0.2
ffmpy==0.6.1
filelock==3.18.0
Flask==3.1.1
gunicorn
cloudinary==1.41.0
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.59.0
fpdf==1.7.2
frappe-bench==5.25.9
frozenlist==1.7.0
fsspec==2025.7.0
git-filter-repo==2.47.0
gitdb==4.0.12
giteasy==10.2.2
githubdata==16.1.2
gitignore_parser==0.1.13
GitPython==3.1.45
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
groovy==0.1.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.7
honcho==2.0.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
language_data==1.3.0
llvmlite==0.44.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mirutil==25.2.0
more-itertools==10.7.0
mpmath==1.3.0
msgspec==0.19.0
mtok==5.0.0
multidict==6.6.4
multiprocess==0.70.18
murmurhash==1.0.13
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.5
oauthlib==3.3.1
onnxruntime==1.23.0
openai==1.109.1
openai-whisper==20250625
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
parse==1.20.2
parso==0.8.5
persiantools==5.3.0
pexpect==4.9.0
pillow==10.4.0
plotly==6.3.0
plum-dispatch==2.5.7
preshed==3.0.10
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
purl==1.6
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.11.0
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.8.0
PyMuPDF==1.26.4
PyNaCl==1.6.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.9.0
pyppeteer==2.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pyquery==2.0.1
pytesseract==0.3.13
python-crontab==2.6.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rapidocr-onnxruntime==1.2.3
referencing==0.36.2
regex==2025.7.34
reportlab==4.4.3
requests==2.32.5
requests-html==0.10.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
rpds-py==0.27.0
rsa==4.9.1
ruff==0.12.8
run_py==3.0.0
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
semantic-version==2.10.0
sentry-sdk==2.42.0
setuptools==80.7.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
stack-data==0.6.3
starlette==0.37.2
streamlit==1.49.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
uv==0.7.22
uvicorn==0.30.3
uvloop==0.21.0
w3lib==2.3.1
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xlsxwriter==3.2.5
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.54.0
httpx==0.27.0
pillow==10.1.0
pydantic==2.5.0
pymupdf==1.23.8
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
gunicorn==21.2.0